+++ test/dart_api/function_test.dart
@@ -12,19 +12,19 @@ import 'package:sass/src/exception.dart';
 
 void main() {
   test(
-      "new Callable() throws a SassFormatException if the argument list is "
-      "invalid", () {
+      "Callable.function() throws a SassFormatException if the argument list "
+      "is invalid", () {
     expect(
-      () => Callable("foo", "arg", (_) => sassNull),
+      () => Callable.function("foo", "arg", (_) => sassNull),
       throwsA(const TypeMatcher<SassFormatException>()),
     );
   });
 
   test(
-      "new AsyncCallable() throws a SassFormatException if the argument list "
-      "is invalid", () {
+      "AsyncCallable.function() throws a SassFormatException if the argument "
+      "list is invalid", () {
     expect(
-      () => AsyncCallable("foo", "arg", (_) async => sassNull),
+      () => AsyncCallable.function("foo", "arg", (_) async => sassNull),
       throwsA(const TypeMatcher<SassFormatException>()),
     );
   });
@@ -33,7 +33,7 @@ void main() {
     var result = compileString(
       'a {b: foo(bar)}',
       functions: [
-        Callable(
+        Callable.function(
           "foo",
           r"$arg",
           expectAsync1((arguments) {
@@ -52,7 +52,7 @@ void main() {
     var result = await compileStringAsync(
       'a {b: foo(bar)}',
       functions: [
-        AsyncCallable(
+        AsyncCallable.function(
           "foo",
           r"$arg",
           expectAsync1((arguments) async {
@@ -73,7 +73,7 @@ void main() {
       compileString(
         'a {b: foo()}',
         functions: [
-          Callable(
+          Callable.function(
             "foo",
             "",
             expectAsync1((arguments) {
@@ -92,7 +92,7 @@ void main() {
       compileString(
         'a {b: foo(x, y, z)}',
         functions: [
-          Callable(
+          Callable.function(
             "foo",
             r"$arg1, $arg2, $arg3",
             expectAsync1((arguments) {
@@ -113,7 +113,7 @@ void main() {
     expect(() {
       compileString(
         'a {b: foo()}',
-        functions: [Callable("foo", "", (arguments) => throw "heck")],
+        functions: [Callable.function("foo", "", (arguments) => throw "heck")],
       );
     }, throwsA(const TypeMatcher<SassException>()));
   });
@@ -122,7 +122,7 @@ void main() {
     var result = compileString(
       'a {b: foo()}',
       functions: [
-        Callable(
+        Callable.function(
           "foo",
           r"$arg: 1",
           expectAsync1((arguments) {
@@ -141,7 +141,7 @@ void main() {
     var result = compileString(
       'a {b: foo(1, 2, 3)}',
       functions: [
-        Callable(
+        Callable.function(
           "foo",
           r"$args...",
           expectAsync1((arguments) {
@@ -164,7 +164,7 @@ void main() {
     var result = compileString(
       r'a {b: foo($bar: 1)}',
       functions: [
-        Callable(
+        Callable.function(
           "foo",
           r"$args...",
           expectAsync1((arguments) {
@@ -188,7 +188,7 @@ void main() {
         compileString(
           'a {b: foo_bar()}',
           functions: [
-            Callable(
+            Callable.function(
               "foo-bar",
               "",
               expectAsync1((arguments) {
@@ -207,7 +207,7 @@ void main() {
         compileString(
           'a {b: foo-bar()}',
           functions: [
-            Callable(
+            Callable.function(
               "foo_bar",
               "",
               expectAsync1((arguments) {