+++ test/dart_api/value/function_test.dart
@@ -43,9 +43,9 @@ void main() {
     var result = compileString(
       "a {b: call(foo(), 12)}",
       functions: [
-        Callable("foo", "", (_) {
+        Callable.function("foo", "", (_) {
           return SassFunction(
-            Callable(
+            Callable.function(
               "bar",
               r"$arg",
               (arguments) => SassNumber(arguments[0].assertNumber().value + 1),